requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyMySQL==1.1.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
openpyxl==3.1.2
pandas>=2.0.0
gunicorn==21.2.0
email-validator==2.2.0
